Compatibility and performance updates
Removed the portscanner package, because port-scanning is extremely slow (at least on Windows, didn't test on other platforms).

As a replacement, a custom port scanner has been added. The new port scanner should work on all platforms since it relies on the `net` module to test if binding to the port works. If it does not, and the error message is 'EADDRINUSE', the port is currently in use. That should be a very reliable scan, and it is pretty fast.

The `getWebHelperLocation` function has been changed to throw an error in case the web helper can't be found, except simply returning false.

A small hack has been added to the `webHelperRunning` function, to prevent a 'windows not supported' error on Windows only, that happens when attempting to use `ps.lookup`. It would probably be wise to look into this some more, since comprehensive windows support is claimed by node-ps.

Furthermore, the spotilocal protocol changed from https to http in recent Spotify versions (https doesn't seem to be supported anymore). This commit changes the protocol in code to reflect those changes.